chore: add snapshot data to debug page
