+++ src/pages/debug.css
@@ -1,3 +1,7 @@
 .DebugPage__Submit {
   padding-bottom: 32px;
 }
+
+.DebugPage__Section {
+  padding-bottom: 32px;
+}
